refactor(reposicao): split AdminReposicaoAplicacao (656→172)
Extrai a lógica (3 queries, filteredItens, 4 mutations, handleAplicarLote/
toggleAll) para o hook useAplicacaoFila e quebra o JSX em AplicacaoHeader,
ProntosTab, InativosTab, AplicadosTab e AplicacaoConfirmDialogs (lote +
individual). Reusa DeltaArrow/SubstituicaoPendenteCard/SubstituicaoModal já
extraídos. Comportamento 1:1; +17 testes.
